feat: handle date questions and set published signals as complete
