+++ src/webapp/components/questionnaire/QuestionInput.tsx
@@ -6,6 +6,7 @@ import NumberWidget from "./widgets/NumberWidget";
 import SingleSelect from "./widgets/SingleSelectWidget";
 import TextWidget from "./widgets/TextWidget";
 import YesNoWidget from "./widgets/YesNoWidget";
+import DatePickerWidget from "./widgets/DatePickerWidget";
 
 export interface QuestionWidgetProps {
     onChange: (question: Question) => void;
@@ -54,6 +55,15 @@ export const QuestionWidget: React.FC<QuestionWidgetProps> = React.memo(props =>
                     onChange={value => onChange(update(question, value))}
                     disabled={disabled}
                     multiline={question.multiline}
+                />
+            );
+
+        case "date":
+            return (
+                <DatePickerWidget
+                    value={question.value}
+                    onChange={value => onChange(update(question, value))}
+                    disabled={disabled}
                 />
             );
         default:
